+++ src/plugins/data/server/index.ts
@@ -145,6 +145,7 @@ export {
   FieldDescriptor as IndexPatternFieldDescriptor,
   shouldReadFieldFromDocValues, // used only in logstash_fields fixture
   FieldDescriptor,
+  decideLegacyClient,
 } from './index_patterns';
 
 export {