+++ src/plugins/vis_type_timeseries/server/lib/vis_data/helpers/get_index_pattern.d.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { ReqFacade } from '../../search_strategies/strategies/abstract_search_strategy';
+
+export function getIndexPatternObject(
+  requestContext: ReqFacade,
+  indexPattern: string
+): Promise<{
+  indexPatternObject: any;
+  indexPatternString: string;
+  dataSourceId?: string;
+}>;